requirements.txt
fastapi
uvicorn
streamlit
gspread
google-auth
google-auth-oauthlib
google-auth-httplib2
openai
pdfplumber
pytesseract
Pillow
python-dotenv
sentry-sdk
pandas
PyYAML
httpx
requests
anthropic
google-genai 